pyproject.toml
[tool.poetry]
name = "rtsp-timelapse"
version = "0.1.0"
description = ""
authors = ["Yannick Linke <invisi@0x0f.net>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
apprise = "^1.7.6"
ffmpeg-python = "^0.2.0"
pydantic = "^2.7.1"
pydantic-settings = "^2.2.1"
schedule = "^1.2.1"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
mypy = "~1.10.0"

[tool.mypy]
plugins = ["pydantic.mypy"]
ignore_missing_imports = true
follow_imports = "silent"
warn_redundant_casts = true
warn_unused_ignores = true
check_untyped_defs = true
no_implicit_reexport = true
disallow_untyped_defs = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
